+++ docs/conf.py
@@ -18,6 +18,7 @@ extensions = [
     "sphinxcontrib.log_cabinet",
     "sphinx_issues",
 ]
+autodoc_typehints = "description"
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "markupsafe": ("https://markupsafe.palletsprojects.com/", None),