+++ CHANGES.rst
@@ -72,6 +72,12 @@ Unreleased
     -   ``pass_environment`` replaces ``environmentfunction`` and
         ``environmentfilter``.
 
+-   Async support no longer requires Jinja to patch itself. It must
+    still be enabled with ``Environment(enable_async=True)``.
+    :issue:`1390`
+-   Overriding ``Context.resolve`` is deprecated, override
+    ``resolve_or_missing`` instead. :issue:`1380`
+
 
 Version 2.11.3
 --------------